Moves parse_dicom
