+++ tests/dicom-parse/common.cpp
@@ -1,29 +1,7 @@
-#define DEBUG_
 #include <dicom-element.h>
 #include <sstream>
 #include <iostream>
-#include <cassert>
 #include <vector>
-
-void parse_dicom(const char* hex_buffer, const char* buffer, size_t size) {
-    printf("parsing dicom, size: %zu Bytes\n",size);
-    char msg_buffer[256] = {0};
-    size_t preamble = 128;
-    size_t prefix = 4;
-    printf("check size\n");
-    assert(size > preamble + prefix);
-    printf("check DICM\n");
-    std::cout << std::string(std::string_view(buffer+preamble,prefix)) << std::endl;
-    assert(std::string_view(buffer+preamble,prefix) == "DICM");
-    // move the read head to where the tag data begins
-    //buffer = buffer+preamble+prefix;
-    printf("read tags\n");
-    for(size_t i = preamble+prefix; i < size;){
-        DicomElement element(buffer, i, hex_buffer);
-        //printf("parse_dicom: bytes: %zu (%zu)\n", element.bytes, element.idx);
-        i = element.GetNextIndex();
-    }
-}
 
 char* hexbuf_to_bytebuf(std::string hex){
     std::stringstream ss;
