+++ tests/dicom-parse/dp-main.cpp
@@ -3,11 +3,12 @@
 #include <filesystem>
 #include <fstream>
 #include <iostream>
+#include <cassert>
 
 namespace fs = std::filesystem;
 
 extern char* bytebuf_to_hexbuf(std::string bytes);
-extern void parse_dicom(const char* hex_buffer, const char* buffer, size_t size);
+void parse_dicom(const char* hex_buffer, const char* buffer, size_t size);
 //uint32_t read_tag_code(std::string hex);
 
 int main(int argc, char **argv) {
@@ -25,4 +26,24 @@ int main(int argc, char **argv) {
     delete[] hex_buffer;
     delete[] buffer;
     return 0;
+}
+
+void parse_dicom(const char* hex_buffer, const char* buffer, size_t size) {
+    printf("parsing dicom, size: %zu Bytes\n",size);
+    char msg_buffer[256] = {0};
+    size_t preamble = 128;
+    size_t prefix = 4;
+    printf("check size\n");
+    assert(size > preamble + prefix);
+    printf("check DICM\n");
+    std::cout << std::string(std::string_view(buffer+preamble,prefix)) << std::endl;
+    assert(std::string_view(buffer+preamble,prefix) == "DICM");
+    // move the read head to where the tag data begins
+    //buffer = buffer+preamble+prefix;
+    printf("read tags\n");
+    for(size_t i = preamble+prefix; i < size;){
+        DicomElement element(buffer, i, hex_buffer);
+        //printf("parse_dicom: bytes: %zu (%zu)\n", element.bytes, element.idx);
+        i = element.GetNextIndex();
+    }
 }
